requirements.txt
selenium>=4.10.0
webdriver-manager>=3.8.6
streamlit>=1.10.0
google-generativeai>=0.1.0
pandas>=1.3.0
python-dotenv>=0.19.0